backend/requirements.txt
# --- Core API ---
fastapi==0.111.0
uvicorn[standard]==0.30.1

# --- Data validation & settings ---
pydantic==2.7.4
pydantic-settings==2.3.3

# --- Database ORM ---
sqlalchemy==2.0.32
alembic==1.13.2   # for migrations (optional but recommended)
#----AI----
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
joblib==1.2.0

# --- Authentication ---
python-jose==3.3.0     # JWT tokens
passlib[bcrypt]==1.7.4 # password hashing

# --- Database drivers ---
psycopg[binary]==3.2.1   # for Postgres (not needed if only SQLite)
# SQLite works without extra drivers

# --- HTTP requests (AI service calls, etc.) ---
httpx==0.27.0

# --- Blockchain (web3) ---
web3==6.19.0
itsdangerous>=2.1.2

#-----Google-----
google-auth==2.29.0

#------Stripe----
stripe>=10,<11
